+++ ui/src/mixins/index.js
@@ -1,19 +1,21 @@
 import contextMixin from './context.js'
 import formMixin from './form.js'
 import generatorMixin from './generator.js'
+import mapMarkersMixin from './map-markers.js'
 import passwordMixin from './password.js'
 import screenMixin from './screen.js'
 import unsavedChangesMixin from './unsaved-changes.js'
+import uploaderMixin from './uploader.js'
 import viewMixin from './view.js'
-import mapMarkersMixin from './map-markers.js'
 
 export {
   contextMixin,
   formMixin,
   generatorMixin,
+  mapMarkersMixin,
   passwordMixin,
   screenMixin,
   unsavedChangesMixin,
-  viewMixin,
-  mapMarkersMixin
+  uploaderMixin,
+  viewMixin
 }