+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `QasUploader` new props: (acceptResizeTypes, picaResizeOptions, sizeLimit, useResize).
 
 ### Changed
-- Changed slot name `tab-slot-[tab.label]` to `tab-after[tab.value]`.
+- Changed slot name `tab-slot-[tab.label]` to `tab-after-[tab.value]`.
 - Changed slot `tab-[tab.label]` to `tab-[tab.value]`.
 - Updated `vue` and `vue-template-compiler` to `v2.6.14` for fixing storybook error.